+++ R/rmGeoIdentity.R
@@ -1,12 +1,22 @@
 #' Remove geographical identity
+#' 
 #' Build a buffer around the a set of geographical coordinates 
-#'  and take a random point around the buffer. The fuction is can 
+#'  and take a random point around the buffer. The function is 
 #'  used to omit the precise location of tricot participants 
 #'  but keeping a close distance to its agro-environment 
 #' @param lonlat a data.frame or matrix with geographical coordinates long lat
-#' @param dist numeric, buffer distance for all \code{lonlat}
+#' @param dist numeric, buffer distance for all \var{lonlat}
 #' @param nQuadSegs integer, number of segments per quadrant
 #' @param ... further arguments passed to \code{\link[sf]{sf}} methods
+#' @examples 
+#' xy <- matrix(c(11.097799, 60.801090,
+#'                11.161298, 60.804199,
+#'                11.254428, 60.822457),
+#'              nrow = 3, ncol = 2, byrow = TRUE)
+#' rmGeoIndentity(xy)
+#'  
+#' @importFrom sf st_point st_sfc st_buffer st_sample
+#' @export
 rmGeoIndentity <- function(lonlat, dist = 0.015, nQuadSegs = 2L, ...){
   
   lonlat <- as.matrix(lonlat)
@@ -26,7 +36,8 @@ rmGeoIndentity <- function(lonlat, dist = 0.015, nQuadSegs = 2L, ...){
   # set the buffer around the points
   lonlatb <- sf::st_buffer(lonlat,
                            dist = dist,
-                           nQuadSegs = nQuadSegs)
+                           nQuadSegs = nQuadSegs, 
+                           ...)
   
   result <- split(lonlatb, seq_len(n))