+++ NEWS.md
@@ -4,6 +4,8 @@ ClimMobTools 0.4.1 (2021-10-10)
 ### IMPROVEMENTS
 
 * `randomise()` deals with unbalanced proportions of technologies, meaning that we can test a whole set of technologies even when not all have equal availability 
+* `rmGeoIdentity()` is added to offer an approach to remove geographical identity of participants
+* `rankTricot()` builds a PlackettLuce rankings using the tricot data 
 
 
 ClimMobTools 0.3.9 (2021-04-12)